% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hist.russian}
\alias{hist.russian}
\title{Lexical extinction data (Russian)}
\format{A matrix with 50,072 rows and 8 columns:
\describe{
  \item{\code{word}}{the item in the word naming task}
  \item{\code{time}}{time of the last observation; this equal 2000 if the word remained in the language and the decade in which the word disappeared from the language otherwise}
  \item{\code{status}}{status of the word; 0 if the word remained in the language, 1 if the word disappeared from the language}
  \item{\code{log.frequency}}{the (log-transformed) frequency of the word in the Google n-gram data for the decade from 1800 to 1810}
  \item{\code{sqrt.length}}{the (square root transformed) length of the word in letters}
  \item{\code{log.old}}{the (log-transformed) orthographic Levenshtein distance between the word and its 5 closest orthographic neighbors in 1810}
  \item{\code{snd}}{the average semantic similarity between the word and its 5 closest semantic neighbors in 1810}
  \item{\code{osc}}{the consistency of the mapping from orhography to semantics in 1810}
}}
\source{
Michel, J. B., Shen, Y. K., Aiden, A. P., Veres, A., Gray, M. K., Pickett, J. P., ... & Pinker, S. (2011). 
Quantitative analysis of culture using millions of digitized books. Science, 331(6014), 176-182.
}
\usage{
hist.russian
}
\description{
Lexical extinction data for Russian based on the Google books n-gram data (Michel et al., 2011)
}
\keyword{datasets}
