% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ld.chin}
\alias{ld.chin}
\title{Chinese lexical decision data}
\format{A matrix with 9,602 rows and 7 columns:
\describe{
  \item{\code{word}}{the item in the lexical decision task}
  \item{\code{rt}}{the average response time in ms}
  \item{\code{log.frequency}}{the (log-transformed) frequency of the word in the CLD}
  \item{\code{length}}{the length of the word in characters}
  \item{\code{sqrt.strokes}}{the (square root transformed) of the number of strokes in the word as a whole}
  \item{\code{log.nwf}}{the (log-transformed) average of the number of words the characters in the word appear in}
  \item{\code{snd}}{the average semantic similarity between the word and its 5 closest semantic neighbors}
}}
\source{
Tsang, Y. K., Huang, J., Lui, M., Xue, M., Chan, Y. W. F., Wang, S., & Chen, H. C. (2018). MELD-SCH: A megastudy of lexical decision in simplified Chinese. Behavior research methods, 50(5), 1763-1777.
}
\usage{
ld.chin
}
\description{
Lexical decision data for Mandarin Chinese from Tsang et al. (2018)
}
\references{
Sun, C. C., Hendrix, P., Ma, J., & Baayen, R. H. (2018). Chinese lexical database (CLD). Behavior Research Methods, 50(6), 2606-2629.
}
\keyword{datasets}
