% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nam}
\alias{nam}
\title{Word naming data (aging)}
\format{A matrix with 4,422 rows and 8 columns:
\describe{
  \item{\code{word}}{the item in the word naming task}
  \item{\code{rt}}{the average response time in ms}
  \item{\code{age}}{the age of the participants}
  \item{\code{log.frequency}}{the (log-transformed) frequency of the word in the SUBTLEX-US corpus}
  \item{\code{length}}{the length of the word in letters}
  \item{\code{log.old}}{the (log-transformed) orthographic Levenshtein distance between the word and its 20 closest orthographic neighbors}
  \item{\code{snd}}{the average semantic similarity between the word and its 5 closest semantic neighbors}
  \item{\code{pos}}{the dominant parts-of-speech category for the word}
}}
\source{
Spieler D. H., & Balota, D. A. (1997). Bringing computational models of word 
naming down to the item level. Psychological Science, 8(6), 411-416.
}
\usage{
nam
}
\description{
Word naming data for old and young participants from Spieler and Balota (1997)
}
\references{
Brysbaert, M., New, B., & Keuleers, E. (2012). Adding part-of-speech information to the 
SUBTLEX-US word frequencies. Behavior Research Methods, 44(4), 991-997.
}
\keyword{datasets}
