% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sd}
\alias{sd}
\title{Semantic decision data}
\format{A matrix with 4,422 rows and 8 columns:
\describe{
  \item{\code{word}}{the item in the semantic decision task}
  \item{\code{rt}}{the average response time in ms}
  \item{\code{log.frequency}}{the (log-transformed) frequency of the word in the SUBTLEX-US corpus}
  \item{\code{length}}{the length of the word in letters}
  \item{\code{log.old}}{the (log-transformed) orthographic Levenshtein distance between the word and its 20 closest orthographic neighbors}
  \item{\code{snd}}{the average semantic similarity between the word and its 5 closest semantic neighbors}
  \item{\code{pos}}{the dominant parts-of-speech category for the word}
  \item{\code{type}}{the semantic type of the word; concrete or abstract}
  \item{\code{concrete.rating}}{the concreteness rating of the word}
}}
\source{
Pexman, P. M., Heard, A., Lloyd, E., & Yap, M. J. (2017). The Calgary semantic decision project: 
concrete/abstract decision data for 10,000 English words. Behavior Research Methods, 49(2), 407-417.
}
\usage{
sd
}
\description{
Semantic decision data (concrete/abstract) from the Calgary semantic decision project
}
\references{
Brysbaert, M., New, B., & Keuleers, E. (2012). Adding part-of-speech information to the 
SUBTLEX-US word frequencies. Behavior Research Methods, 44(4), 991-997.
}
\keyword{datasets}
