% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vb}
\alias{vb}
\title{Paste tense generation}
\format{A matrix with 1,978 rows and 7 columns:
\describe{
  \item{\code{word}}{the item in the paste tense generation task}
  \item{\code{rt}}{the average response time in ms}
  \item{\code{rt}}{regularity of the verb}
  \item{\code{log.frequency}}{the (log-transformed) frequency of the word in the SUBTLEX-US corpus}
  \item{\code{length}}{the length of the word in letters}
  \item{\code{log.old}}{the (log-transformed) orthographic Levenshtein distance between the word and its 20 closest orthographic neighbors}
  \item{\code{snd}}{the average semantic similarity between the word and its 5 closest semantic neighbors}
  \item{\code{pos}}{the dominant parts-of-speech category for the word}
  \item{\code{type}}{the semantic type of the word; concrete or abstract}
  \item{\code{concrete.rating}}{the concreteness rating of the word}
}}
\source{
Cohen-Shikora, E. R., Balota, D. A., Kapuria, A., & Yap, M. J. (2013). The past tense inflection 
project (PTIP): Speeded past tense inflections, imageability ratings, and past tense consistency measures 
for 2,200 verbs. Behavior Research Methods, 45(1), 151-159.
}
\usage{
vb
}
\description{
Past tense generation data from Cohen et al. (2013)
}
\references{
Brysbaert, M., New, B., & Keuleers, E. (2012). Adding part-of-speech information to the 
SUBTLEX-US word frequencies. Behavior Research Methods, 44(4), 991-997.
}
\keyword{datasets}
